% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{parse_consecutive_ids}
\alias{parse_consecutive_ids}
\title{Parce consecutive ids}
\usage{
parse_consecutive_ids(id)
}
\arguments{
\item{input}{The id
DO NOT REMOVE.}
}
\value{
array with ids
}
\description{
Function to split string with consecutive ids into an array (the separator is -)
}
