% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{parse_multiple_ids}
\alias{parse_multiple_ids}
\title{Parse multiple ids}
\usage{
parse_multiple_ids(id)
}
\arguments{
\item{id}{The id
DO NOT REMOVE.}
}
\value{
array with ids
}
\description{
Function to split a string with multiple ids into an array, separators can be - or ,
}
