% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{process_biosamples}
\alias{process_biosamples}
\title{Process biosamples}
\usage{
process_biosamples(input)
}
\arguments{
\item{input}{The Biosample id
DO NOT REMOVE.}
}
\value{
dataframe with biosamples
}
\description{
Function to retrieve BioSamples metadata and create a dataframe from it, also writing a csv file
}
