% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{process_study}
\alias{process_study}
\title{Process ENA Study or Project}
\usage{
process_study(input, xml)
}
\arguments{
\item{input}{The ENA Study or Project id}

\item{xml}{The jats skeleton xml
DO NOT REMOVE.}
}
\value{
string with html content
}
\description{
Function to retrieve ENA metadata and BioSamples and ArrayExpress identifiers, import them into the relevant fields of the data paper template concept and generate HTML visualisation
}
